+++ README.md
@@ -1,2 +1,2 @@
 # ProductivityPlanner
-Personal project to build a productivity planner application
+Personal project to build a productivity planner domain
