+++ TaskService/src/main/java/TaskServiceApplication.java
@@ -3,7 +3,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.ComponentScan;
 
 @SpringBootApplication
-@ComponentScan({"configuration", "controller"})
+@ComponentScan({"configurations", "controllers"})
 public class TaskServiceApplication {
 
 	public static void main(String[] args) {
