+++ TaskService/src/main/java/domain/ITaskService.java
@@ -1,4 +1,4 @@
-package application;
+package domain;
 
 import java.util.List;
