+++ TaskService/src/main/java/domain/entities/ProductivityTask.java
@@ -1,6 +1,4 @@
-package application;
-
-import java.util.Date;
+package domain.entities;
 
 public class ProductivityTask {
