+++ TaskService/src/main/java/repositories/mysql/IMysqlConfiguration.java
@@ -1,4 +1,4 @@
-package repository.mysql;
+package repositories.mysql;
 
 public interface IMysqlConfiguration {
